fix: remove JDK 17 hard requirement and break circular build dep
- recent PDK update means JDK 17+ is supported
- decouble some basic Bedrock related IDs until that is moved in galileo-sdk
